Refactor clasp pull
